Rename NewExp to NewExperiment. Extract ExpLog to ExperimentListing component. Start implementing the typescript experiment data types.
